+++ src/dedede/frame.rs
@@ -13,3 +13,21 @@ use smash::app::*;
 use smash::phx::Vector3f;
 use crate::util::*;
 use super::*;
+pub fn install() {
+	smashline::install_agent_frames!(dedede_frame);
+}
+#[fighter_frame(agent = FIGHTER_KIND_DEDEDE)]
+pub fn dedede_frame(fighter : &mut L2CFighterCommon) {
+    unsafe {
+		let boma = smash::app::sv_system::battle_object_module_accessor(fighter.lua_state_agent); 
+		let status_kind = smash::app::lua_bind::StatusModule::status_kind(boma);
+		let ENTRY_ID = WorkModule::get_int(boma, *FIGHTER_INSTANCE_WORK_ID_INT_ENTRY_ID) as usize;
+		let fighter_kind = smash::app::utility::get_kind(boma);
+
+		if [hash40("attack_lw3")].contains(&MotionModule::motion_kind(boma)) {
+			if MotionModule::frame(boma) >= 47.0 {
+				StatusModule::change_status_request_from_script(boma, *FIGHTER_STATUS_KIND_WAIT, true);
+			};
+		};
+	};
+}
